fix: move ConfLogo to components
